Rewrote Spy/Secret instantiation.
Basically, a lot of generic helper methods are simply missing from
BasicObject, such as tap and __method__ and all others, so I just pass the
block from Spy right on to Secrets and it does the work there where it
was going to end up anyway.

I'm also using TOISB for getting info about BasicObject instances as
well.